Updated conditional delete
Updated the way conditional deletes are handled when delete
interceptors are registered. This update tries to take advantage of
Mongo’s bulk delete operation using a list of resource IDs.

Also added additional error checking (to be more thorough) and updated
some of the function comments to better suit godoc.