Fixed delete() and conditionalDelete() operations
The ordering of get/delete operations was mistakenly reversed.
Therefore, interceptors registered for a delete operation would fail
because they would be unable to get the resource that was just deleted.

I also reviewed the conditionalDelete() operation and removed the
return statement that would cause the operation to fail because of the
interceptor. As a design decision we’ve decided that interceptors
should always fail silently and not interrupt the more important
database operation.

Also: Added a simplification to invokeInterceptors() that makes the
code marginally more efficient.