Work with the *gin.Engine after FHIR routes
Users of the server of the library have the ability to add middleware to the
FHIR resources. The way that the API works, users have access to the underlying
*gin.Engine to add any kind of middleware, but anything they do with it will
happen *before* the FHIR routes. This means that any routes registered will not
be passed through the middleware registered by the main FHIR server. This is a
huge problem when it comes to things like security. This commit lets users
register functions that will set up routes *after* the FHIR server is set up.
That allows them to be protected by security, or any other middleware, that the
FHIR server creates.