Added Mongo Indexes and Updated Server Config
I created a mechanism that creates a default set of indexes on the mongo database to improve performance at scale. For now these indexes are only created on referenced ID fields used during  operations. However, there is space in the config file (config/indexes.conf) to create additional indexes if needed.

During testing I ran into some complications with the way the server is configured. Basically, the name of the database used by the server was hard-coded. I updated the way the server is configured to use a DefaultConfig struct that can be modified if needed (e.g. for testing). I also updated the server's tests to reflect these changes (although the tests would have worked the old way too).