Refactored Search
Refactored the CreateQuery and CreatePipline methods into one Search() method that handles all mongo searches. This returns a generic results interface{} as well as a total count of the resources found on the server and any errors encountered. I refactored the tests to remove all references to CreateQuery and CreatePipeline. The decision to use a mongo pipeline for searches is now hidden from the caller, making the searcher more flexible.